fix(split-app-frontend): fixed header bar visibility in user and login page
