+++ components/screenComponents/Header.tsx
@@ -3,44 +3,21 @@
 import React, { useState, useEffect } from "react";
 import Image from "next/image";
 import Link from "next/link";
-import { useRouter } from "next/navigation";
 import { usePathname } from "next/navigation";
-import { useAppStore } from "@/store/zustand";
 import { cn } from "@/lib/utils";
-import {
-  DropdownMenu,
-  DropdownMenuContent,
-  DropdownMenuItem,
-  DropdownMenuTrigger,
-} from "@/components/ui/dropdown-menu";
-import DPrimaryButtonLink from "@/components/DPrimaryButtonLink";
-import useShowToast from "@/components/DToast";
-import { logOut } from "@/lib/authApi";
 
-const authRoutes = ["/login", "/signup"];
+import DPrimaryButtonLink from "@/components/DPrimaryButtonLink";
+
+const headerRoutes = ["/"];
 
 const Header = () => {
   const pathname = usePathname();
-  const showHeader = !authRoutes.includes(pathname);
-  const { user, clearAll } = useAppStore();
-  const router = useRouter();
-  const { showSuccessToast } = useShowToast();
+  const showHeader = headerRoutes.includes(pathname);
   const [isMounted, setIsMounted] = useState(false);
 
   useEffect(() => {
     setIsMounted(true);
   }, []);
-
-  const onLogOut = async () => {
-    try {
-      await logOut();
-      clearAll();
-      router.push("/");
-      showSuccessToast("You have successfully logged out.");
-    } catch (error) {
-      console.log("Logout error", error);
-    }
-  };
 
   return (
     <div
@@ -56,50 +33,20 @@ const Header = () => {
       </div>
 
       {isMounted && (
-        <>
-          {user ? (
-            <div className="flex items-center gap-4">
-              <p>{user.name ?? user.email}</p>
+        <div className="w-full flex gap-8 items-center justify-end flex-1">
+          <Link
+            href="/login"
+            className="primaryFontSize font-semibold cursor-pointer hover:underline"
+          >
+            Log In
+          </Link>
 
-              <DropdownMenu>
-                <DropdownMenuTrigger>
-                  <Image
-                    src="/guy1Profile.png"
-                    height={24}
-                    width={24}
-                    alt="Profile Photo"
-                  />
-                </DropdownMenuTrigger>
-                <DropdownMenuContent className="shadow-lg w-40" align="end">
-                  <DropdownMenuItem className="cursor-pointer">
-                    Your Account
-                  </DropdownMenuItem>
-                  <DropdownMenuItem
-                    className="cursor-pointer"
-                    onClick={onLogOut}
-                  >
-                    Sign Out
-                  </DropdownMenuItem>
-                </DropdownMenuContent>
-              </DropdownMenu>
-            </div>
-          ) : (
-            <div className="w-full flex gap-8 items-center justify-end flex-1">
-              <Link
-                href="/login"
-                className="primaryFontSize font-semibold cursor-pointer hover:underline"
-              >
-                Log In
-              </Link>
-
-              <DPrimaryButtonLink
-                href="/signup"
-                label="Sign Up"
-                className="w-min"
-              />
-            </div>
-          )}
-        </>
+          <DPrimaryButtonLink
+            href="/signup"
+            label="Sign Up"
+            className="w-min"
+          />
+        </div>
       )}
     </div>
   );
